% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_site_skeleton.R
\name{course_site_skeleton}
\alias{course_site_skeleton}
\title{course_site_skeleton}
\usage{
course_site_skeleton(path)
}
\arguments{
\item{path}{The path to be used to generate the templates}
}
\description{
course_site_skeleton
}
