% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_formatted_title.R
\name{make_formatted_title}
\alias{make_formatted_title}
\title{ggplot2 Title Break Make}
\usage{
make_formatted_title(title, character_width = 20)
}
\arguments{
\item{title}{the vector of values you would like to add carriage returns to}

\item{character_width}{how many characters do you want each line to be maximum}
}
\description{
this function puts line breaks into long strings
}
\examples{
a <- "A really long title for a graph that will probably not fit but we need to write it all"

make_formatted_title(a)

}
