% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pres_wake.R
\name{theme_pres_wake}
\alias{theme_pres_wake}
\title{Gggplot2 Theme for Presentations}
\usage{
theme_pres_wake()
}
\description{
This theme is a larger size for presentations consistent with
    the WFU style guide
}
