% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wake.R
\name{theme_wake}
\alias{theme_wake}
\title{Gggplot2 Theme for Wake Forest University}
\usage{
theme_wake()
}
\description{
This theme is a larger size for presentations consistent with
    the WFU style guide
}
